requirements.txt
deepgram-sdk
python-dotenv
jiwer
pandas 
pydub
assemblyai
num2words